refactor(tests): split test_nmap into nmap/ package; share XML fixtures
tests/tools/recon/test_nmap.py (524) split to mirror tools/recon/nmap/:
test_flags.py (TestAssembleFlags), test_parser.py (TestParseXml),
test_scanner.py (TestEvidenceFilename + TestNmapScan). The two real-shape
nmap -oX payloads that parser and scanner tests both leaned on become
shared nmap_xml_two_hosts / nmap_xml_no_hosts fixtures in
tests/fixtures/recon.py - one copy of the scan-result data instead of a
module constant duplicated across the split. All three files <= 207 lines.
